# lib/api-zod/src/generated/types/organizationAiPhishingConsent.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * NovaMail email platform API
 * OpenAPI spec version: 1.0.0
 */
import type { AiPhishingProviderStatus } from './aiPhishingProviderStatus';

export interface OrganizationAiPhishingConsent {
  organizationId: string;
  enabled: boolean;
  /** @nullable */
  consentAt: Date | null;
  provider: AiPhishingProviderStatus;
}